refactor: 커맨드 문구 추가
