+++ src/main/java/bridge/domain/GameCommand.java
@@ -1,20 +1,26 @@
-package bridge.controller;
+package bridge.domain;
 
 import java.util.Arrays;
 import java.util.Objects;
 
 public enum GameCommand {
-    RESTART ("R"),
-    QUIT ("Q");
+    RESTART ("R", "재시도"),
+    QUIT ("Q", "종료");
 
     private final String command;
+    private final String word;
 
-    GameCommand(String command) {
+    GameCommand(String command, String word) {
         this.command = command;
+        this.word = word;
     }
 
     public String getCommand() {
         return command;
+    }
+
+    public String getWord() {
+        return word;
     }
 
     public static GameCommand parseCommandToInstance(String command) {
